kloud: create new terraformtemplate file
