+++ go/src/koding/kites/kloud/kloud/terraformutils.go
@@ -10,7 +10,6 @@ import (
 	"koding/kites/kloud/contexthelper/session"
 	"koding/kites/kloud/klient"
 	"koding/kites/kloud/userdata"
-	"reflect"
 	"strconv"
 	"strings"
 	"sync"
@@ -57,29 +56,6 @@ type terraformCredential struct {
 	Provider   string
 	Identifier string
 	Data       map[string]string `mapstructure:"data"`
-}
-
-type terraformTemplate struct {
-	Resource struct {
-		Aws_Instance map[string]map[string]interface{} `json:"aws_instance"`
-	} `json:"resource"`
-	Provider struct {
-		Aws struct {
-			Region    string `json:"region"`
-			AccessKey string `json:"access_key"`
-			SecretKey string `json:"secret_key"`
-		} `json:"aws"`
-	} `json:"provider"`
-	Variable map[string]map[string]interface{} `json:"variable,omitempty"`
-}
-
-func newTerraformTemplate(content string) (*terraformTemplate, error) {
-	var template *terraformTemplate
-	if err := json.Unmarshal([]byte(content), &template); err != nil {
-		return nil, err
-	}
-
-	return template, nil
 }
 
 func (m *Machines) AppendRegion(region string) {
@@ -304,85 +280,6 @@ func parseProviderAndLabel(resource string) (string, string, error) {
 	return provider, label, nil
 }
 
-func (t *terraformTemplate) injectCustomVariables(data *terraformData) {
-	for _, cred := range data.Creds {
-		prefix := cred.Provider
-		for key, val := range cred.Data {
-			varName := fmt.Sprintf("%s_%s_%s", prefix, cred, key)
-			t.Variable[varName] = map[string]interface{}{
-				"default": val,
-			}
-		}
-	}
-}
-
-func (t *terraformTemplate) injectKodingVariables(data *terraformData) {
-	var properties = []struct {
-		collection string
-		fieldToAdd map[string]bool
-	}{
-		{"User",
-			map[string]bool{
-				"username": true,
-				"email":    true,
-			},
-		},
-		{"Account",
-			map[string]bool{
-				"profile": true,
-			},
-		},
-		{"Group",
-			map[string]bool{
-				"title": true,
-				"slug":  true,
-			},
-		},
-	}
-
-	for _, p := range properties {
-		model, ok := structs.New(data).FieldOk(p.collection)
-		if !ok {
-			continue
-		}
-
-		for _, field := range model.Fields() {
-			fieldName := strings.ToLower(field.Name())
-			// check if the user set a field tag
-			if field.Tag("bson") != "" {
-				fieldName = field.Tag("bson")
-			}
-
-			exists := p.fieldToAdd[fieldName]
-
-			// we need to declare to call it recursively
-			var addVariable func(*structs.Field, string, bool)
-
-			addVariable = func(field *structs.Field, varName string, allow bool) {
-				if !allow {
-					return
-				}
-
-				// nested structs, call again
-				if field.Kind() == reflect.Struct {
-					for _, f := range field.Fields() {
-						newName := varName + "_" + strings.ToLower(f.Name())
-						addVariable(f, newName, true)
-					}
-					return
-				}
-
-				t.Variable[varName] = map[string]interface{}{
-					"default": field.Value(),
-				}
-			}
-
-			varName := "koding_" + strings.ToLower(p.collection) + "_" + fieldName
-			addVariable(field, varName, exists)
-		}
-	}
-}
-
 func injectKodingData(ctx context.Context, content, username string, data *terraformData) (*buildData, error) {
 	sess, ok := session.FromContext(ctx)
 	if !ok {
@@ -411,6 +308,9 @@ func injectKodingData(ctx context.Context, content, username string, data *terra
 
 	// inject koding variables, in the form of koding_user_foo, koding_group_name, etc..
 	template.injectKodingVariables(data)
+
+	// // inject all other variables from credentials
+	// template.injectCustomVariables(data)
 
 	if len(template.Resource.Aws_Instance) == 0 {
 		return nil, fmt.Errorf("instance is empty: %v", template.Resource.Aws_Instance)
@@ -508,13 +408,13 @@ func injectKodingData(ctx context.Context, content, username string, data *terra
 		template.Resource.Aws_Instance[resourceName] = instance
 	}
 
-	out, err := json.MarshalIndent(template, "", "  ")
+	out, err := template.jsonOutput()
 	if err != nil {
 		return nil, err
 	}
 
 	b := &buildData{
-		Template: string(out),
+		Template: out,
 		KiteIds:  kiteIds,
 		Region:   template.Provider.Aws.Region,
 	}
